Refactor to allow extensions to the algebraic simplifier
cherry-pick form: 90355d2af1f1feea996d6ff4ae1fe2e183f5db3e
